Rename internal registry service pkg
