+++ packer/run_interfaces.go
@@ -3,7 +3,7 @@ package packer
 import (
 	"github.com/hashicorp/hcl/v2"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
-	packerregistry "github.com/hashicorp/packer/internal/packer_registry"
+	packerregistry "github.com/hashicorp/packer/internal/registry"
 	plugingetter "github.com/hashicorp/packer/packer/plugin-getter"
 )
